services/knowledge_base/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-generativeai>=0.8.0
pgvector>=0.3.0
python-docx>=1.1.0
pymupdf>=1.25.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
